pyproject.toml
[project]
name = "smareverse-vending-machine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "imagehash>=4.3.2",
    "networkx>=3.5",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.163",
]
